+++ R/package.R
@@ -4,6 +4,7 @@
 #' @importFrom utils head head.matrix tail tail.matrix
 #' @importFrom stats setNames na.omit prop.test binom.test relevel quantile
 #' @importFrom htmltools tags tagList
+#' @importFrom lifecycle deprecated
 #' @importFrom magrittr %>%
 #' @import methods
 NULL